pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'

[tool.poetry]
name = "situtational-2"
version = "0.1.0"
description = ""
authors = ["Bolot Zhusupekov <bolot.zhusupekov07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.3"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-cacheops = "^7.0.2"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
